iOS: rename to device_implementation to make dev easier
