+++ lib/calabash/ios/device/device_implementation.rb
